feat: Move limit methods to `LimitUtil` class.
